i forgot this

+++ yolo/yolo_slowfast.py
@@ -189,7 +189,7 @@ class YOLOStream:
     ):
         for i, (im, pred) in enumerate(zip(yolo_preds.ims, yolo_preds.pred)):
             if args == "TDOA":
-                im = cv2.cvtColor(vision_frame, cv2.COLOR_BGR2RGB)
+                im = cv2.cvtColor(np.float32(vision_frame), cv2.COLOR_BGR2RGB)
 
             else:
                 im = cv2.cvtColor(im, cv2.COLOR_BGR2RGB)
